+++ FF1Lib/TileSetsUpdates.cs
@@ -38,9 +38,5 @@ namespace FF1Lib
 			const byte specialMask = 0b1110_0001;
 			tileSets[(int)TileSets.Castle].Tiles[0x61].PropertyType &= specialMask;
 		}
-
-
-
-
 	}
 }
